external-scada/scada-api/requirements.txt
paho-mqtt==1.6.1
flask==3.0.0
cryptography==42.0.5
